refactor: move tree pre- and postprocessing out of algorithm
